LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_block_modals-3-expected.txt

CONSOLE MESSAGE: Use of window.prompt is not allowed in a sandboxed frame when the allow-modals flag is not set.


PASS Frames without `allow-modals` should not be able to open modal dialogs
